services/orchestrator/requirements.txt
python-telegram-bot>=21.0,<22.0
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.40.0
